fix: bundling on window
we're using webpack and typescript to bundle out application via
electron-forge.  we're using webmidi.js 3 with imports jazz-midi
in a main/node.js process. When bundling on windows using bash
and node 22.x, I am getting the error that it is unable to resolve
module linuxa7/node.  this commit modifies the index.js to more
thoroughly build the path to the native module.